// src/common/types.ts
export interface PrefixSuffix {
	prefix: string
	suffix: string
}

export interface StreamResponse {
	model: string
	created_at: string
	response: string
	content: string
	done: boolean
	context: number[]
	total_duration: number
	load_duration: number
	prompt_eval_count: number
	prompt_eval_duration: number
	eval_count: number
	eval_duration: number
	choices: [
		{
			text: string
			delta: {
				content: string
			}
		},
	]
}
// ...
export interface FimTemplateData extends Record<string, string | undefined> {
	context: string
	fileName: string
	prefix: string
	suffix: string
	systemMessage: string
}
// ...
export interface StreamRequestOptions {
	hostname: string
	path: string
	port: string | number
	protocol: string
	method: string
	headers: Record<string, string>
}
// ...
export type ResolvedInlineCompletion =
	| InlineCompletionItem[]
	| InlineCompletionList
	| PromiseLike<InlineCompletionItem[] | InlineCompletionList | null | undefined>
	| null
	| undefined